+++ CHANGELOG.md
@@ -4,10 +4,15 @@ Tidarr notable changes.
 [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) format.
 
 ## 📦 0.4.5
+### 🚀 Added
+* [Front] A button to download all videos of an artist was added on artist page
 ### 🖍 Changed
 * [API] Run beets only for albums
 * [API] Fix undefined label in notification
 * [Front] Fix console ANSI chars encoding
+* [Front] Update display mode UI
+* [Docker] Remove useless env vars : `ENABLE_PLEX_UPDATE`, `ENABLE_GOTIFY`, `ENABLE_APPRISE_API`
+* [API] Update tiddl to 2.8.0
 
 ## 📦 0.4.4
 ### 🖍 Changed